requirements.txt
# Core Google ADK and AI packages
google-adk>=0.1.0
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.40.0

# LLM and embeddings
litellm>=1.30.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-ollama>=0.1.0

# Document processing
pypdf>=3.17.0
sentence-transformers>=2.2.2
pyarrow>=14.0.0

# Vector store
chromadb>=0.4.22

# Utilities
python-dotenv>=1.0.0

# FastAPI and dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Anthropic Provider
anthropic>=0.39.0

# Google Provider
google-generativeai>=0.8.0